requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
openai==0.27.0
Pillow==8.3.2
requests==2.26.0
